# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRS API",
        "description": "API documentation for the PRS application.",
        "contact": {
            "email": "support@prs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/field-staff/retailer-orders": {
            "get": {
                "tags": [
                    "Field Staff Retailer Orders"
                ],
                "summary": "List retailer orders assigned to the logged-in field staff",
                "operationId": "d4300b1d349722a5ccc18c4847ebf19f",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of assigned retailer orders"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/field-staff/retailer-orders/{id}": {
            "get": {
                "tags": [
                    "Field Staff Retailer Orders"
                ],
                "summary": "Get a single retailer order detail",
                "operationId": "e6443e590080fabe35ea971a3dae3f13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order Details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/field-staff/retailer-orders/{id}/update-status": {
            "post": {
                "tags": [
                    "Field Staff Retailer Orders"
                ],
                "summary": "Accept, reject, or cancel a retailer order",
                "description": "Field Staff can accept (moves to processing), reject, or cancel (if they placed it).",
                "operationId": "4fe4bb96ee7c84cf9d9a1eba44b4f7a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processing",
                                            "rejected",
                                            "cancelled"
                                        ]
                                    },
                                    "cancellation_reason": {
                                        "description": "Required if status is cancelled",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/prescriptions/upload": {
            "post": {
                "tags": [
                    "Prescription"
                ],
                "summary": "Upload a prescription and get OCR results",
                "operationId": "471c4295c9c4fc84bcd9040536d43dc2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "prescription"
                                ],
                                "properties": {
                                    "prescription": {
                                        "description": "The prescription image or PDF",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OCR results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to extract data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to extract data from the prescription."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}